perf: 拆分 shouldPluginEnable  逻辑
