+++ packages/plugin-initial-state/src/utils/index.ts
@@ -1,0 +1,17 @@
+import { utils } from 'umi';
+import { readFileSync } from 'fs';
+// @ts-ignore
+import { init, parse } from 'es-module-lexer/dist/lexer';
+
+export const shouldPluginEnable = async (entryFile?: string) => {
+  let hasExport = false;
+
+  if (entryFile) {
+    await init;
+    const fileContent = readFileSync(entryFile, 'utf-8');
+    const [_, exportsList] = parse(fileContent);
+    hasExport = exportsList.includes('getInitialState');
+  }
+
+  return hasExport;
+};
